+++ lib/services/agecalc.dart
@@ -1,0 +1,7 @@
+class AgeCalc {
+  String calculateAge(DateTime dob) {
+    Duration dur = DateTime.now().difference(dob);
+    final diff = (dur.inDays / 365).floor();
+    return "$diff";
+  }
+}
